backend/requirements.txt
amqp==2.6.1
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.2.10
attrs==20.2.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
black==19.10b0
boto3==1.10.34
botocore==1.13.50
celery==4.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssutils==1.0.2
diff-match-patch==20181111
Django==3.0.6
django-cors-headers==3.2.0
django-extensions==2.2.5
django-filter==2.2.0
django-inbound-email==0.13
django-inlinecss==0.3.0
django-storages==1.8
djangorestframework==3.10.3
# djangorestframework-filters==0.11.1
docutils==0.15.2
drf-extensions==0.6.0
email-reply-parser==0.5.9
factory-boy==2.12.0
Faker==4.1.2
future==0.18.2
graphene==2.1.8
graphene-django==2.7.1
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
jmespath==0.10.0
kombu==4.6.11
lxml==4.5.2
mccabe==0.6.1
pathspec==0.8.0
pdfkit==0.6.1
Pillow==7.2.0
promise==2.3
psycopg2==2.8.4
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pynliner==0.8.0
python-dateutil==2.8.1
python-http-client==3.3.1
pytz==2020.1
quill-delta==1.0.1
redis==3.3.11
regex==2020.7.14
requests==2.24.0
Rx==1.6.1
s3transfer==0.2.1
sendgrid==6.4.6
singledispatch==3.4.0.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
text-unidecode==1.3
toml==0.10.1
twilio==6.45.1
typed-ast==1.4.1
urllib3==1.25.10
vine==1.3.0